2026/MelStudios/MelStudios_WebServer/requirements.txt
fastapi[standard]
pycryptodome
sqlalchemy
gunicorn
faker